* master:
  [core] Fixing squid:S1319 -  Declarations should use Java collection interfaces such as "List" rather than specific implementation classes such as "LinkedList". (manolama - updated bindings added since the PR)
  [core] Use longs instead of ints to support larger key spaces. Changed int to long in Measurements code to support large scale workloads. (manolama - fixed checkstyle errors)
  [core] Export totalHistogram for HdrHistogram measurement
  [core] Add an operation enum to the Workload class. This can eventually be used to replace the strings.
  [core] Add a Fisher-Yates array shuffle to the Utils class.
  [core] Fix an issue where the threadid and threadCount were not passed to the workload client threads. Had to use setters to get around the checkstyle complaint of having too many parameters.
  Upgrading googlebigtable to the latest version. The API used by googlebigtable has had quite a bit of churn.  This is the minimal set of changes required for the upgrade.
  [geode] Update to apache-geode 1.2.0 release
  [core] Update to use newer version of Google Cloud Spanner client and associated required change
  [core] Add a reset() method to the ByteIterator abstract and implementations for each of the children. This lets us re-use byte iterators if we need to access the values again (when applicable).
  [hbase12] Add HBase 1.2+ specific client that relies on the shaded client artifact provided by those versions. (#970)
  [distro] Refresh Apache licence text (#969)
  [memcached] support binary protocol (#965)
  [accumulo] A general "refresh" to the Accumulo binding (#947)
  [cloudspanner] Add binding for Google's Cloud Spanner. (#939)
  [aerospike] Change the write policy to REPLACE_ONLY (#937)

+++ core/src/main/java/com/yahoo/ycsb/generator/AcknowledgedCounterGenerator.java
@@ -31,12 +31,12 @@ public class AcknowledgedCounterGenerator extends CounterGenerator {
 
   private final ReentrantLock lock;
   private final boolean[] window;
-  private volatile int limit;
+  private volatile long limit;
 
   /**
    * Create a counter that starts at countstart.
    */
-  public AcknowledgedCounterGenerator(int countstart) {
+  public AcknowledgedCounterGenerator(long countstart) {
     super(countstart);
     lock = new ReentrantLock();
     window = new boolean[WINDOW_SIZE];
@@ -48,15 +48,15 @@ public class AcknowledgedCounterGenerator extends CounterGenerator {
    * (as opposed to the highest generated counter value).
    */
   @Override
-  public Integer lastValue() {
+  public Long lastValue() {
     return limit;
   }
 
   /**
    * Make a generated counter value available via lastInt().
    */
-  public void acknowledge(int value) {
-    final int currentSlot = (value & WINDOW_MASK);
+  public void acknowledge(long value) {
+    final int currentSlot = (int)(value & WINDOW_MASK);
     if (window[currentSlot]) {
       throw new RuntimeException("Too many unacknowledged insertion keys.");
     }
@@ -68,10 +68,10 @@ public class AcknowledgedCounterGenerator extends CounterGenerator {
       // over to the "limit" variable
       try {
         // Only loop through the entire window at most once.
-        int beforeFirstSlot = (limit & WINDOW_MASK);
-        int index;
+        long beforeFirstSlot = (limit & WINDOW_MASK);
+        long index;
         for (index = limit + 1; index != beforeFirstSlot; ++index) {
-          int slot = (index & WINDOW_MASK);
+          int slot = (int)(index & WINDOW_MASK);
           if (!window[slot]) {
             break;
           }
