+++ core/src/main/java/com/yahoo/ycsb/generator/CounterGenerator.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2010-2016 Yahoo! Inc., 2017 YCSB contributors. All rights reserved.
+ * Copyright (c) 2010 Yahoo! Inc., Copyright (c) 2017 YCSB contributors. All rights reserved.
  * <p>
  * Licensed under the Apache License, Version 2.0 (the "License"); you
  * may not use this file except in compliance with the License. You
@@ -17,29 +17,29 @@
 
 package com.yahoo.ycsb.generator;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Generates a sequence of integers.
  * (0, 1, ...)
  */
 public class CounterGenerator extends NumberGenerator {
-  private final AtomicInteger counter;
+  private final AtomicLong counter;
 
   /**
    * Create a counter that starts at countstart.
    */
-  public CounterGenerator(int countstart) {
-    counter = new AtomicInteger(countstart);
+  public CounterGenerator(long countstart) {
+    counter=new AtomicLong(countstart);
   }
 
   @Override
-  public Integer nextValue() {
+  public Long nextValue() {
     return counter.getAndIncrement();
   }
 
   @Override
-  public Integer lastValue() {
+  public Long lastValue() {
     return counter.get() - 1;
   }
